FEBio2: Moved Properties to FECoreBase. Implemented PIMPL idiom for FEModel. Made FELoadCurve a base class and implemented FEDataLoadCurve and FELinearRamp.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@9284 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FECore/FENewtonSolver.h
@@ -82,7 +82,7 @@ public:
 	// solver parameters
 	int					m_nqnsolver;	//!< quasi-Newton strategy that will be selected
 	int					m_maxups;		//!< max number of quasi-newton updates
-	int					m_cmax;			//!< max condition numbers
+	double				m_cmax;			//!< max condition numbers
 	int					m_maxref;		//!< max nr of reformations per time step
 	FENewtonStrategy*	m_pbfgs;		//!< class handling the specific stiffness update logic
